feat(credential-providers): rename credential chain file
